pyproject.toml
[tool.poetry]
name = "igniteadv"
version = "0.1.0"
description = ""
authors = ["s000ne <sone.h0710@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
pytorch-ignite = "^0.4.2"
torch = "^1.8.1"
torchvision = "^0.9.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
